+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/AADJwtDecoderProviderConfiguration.java
@@ -14,13 +14,14 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 /**
  * Allows resolving configuration from an
- * <a href="https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfig">OpenID Provider Configuration</a> or
- * <a href="https://tools.ietf.org/html/rfc8414#section-3.1">Authorization Server Metadata Request</a> based on provided
- * issuer and method invoked.
+ * <a href="https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfig">OpenID Provider
+ * Configuration</a> or
+ * <a href="https://tools.ietf.org/html/rfc8414#section-3.1">Authorization Server Metadata Request</a> based on
+ * provided issuer and method invoked.
  */
 public class AADJwtDecoderProviderConfiguration {
+
     private static final String OIDC_METADATA_PATH = "/.well-known/openid-configuration";
-    private static final String OAUTH_METADATA_PATH = "/.well-known/oauth-authorization-server";
     private static final RestTemplate rest = new RestTemplate();
     private static final ParameterizedTypeReference<Map<String, Object>> typeReference =
         new ParameterizedTypeReference<Map<String, Object>>() {};
@@ -29,25 +30,8 @@ public class AADJwtDecoderProviderConfiguration {
         return getConfiguration(oidcIssuerLocation, oidc(URI.create(oidcIssuerLocation)));
     }
 
-    public static Map<String, Object> getConfigurationForIssuerLocation(String issuer) {
-        URI uri = URI.create(issuer);
-        return getConfiguration(issuer, oidc(uri), oidcRfc8414(uri), oauth(uri));
-    }
-
-    public static void validateIssuer(Map<String, Object> configuration, String issuer) {
-        String metadataIssuer = "(unavailable)";
-        if (configuration.containsKey("issuer")) {
-            metadataIssuer = configuration.get("issuer").toString();
-        }
-        if (!issuer.equals(metadataIssuer)) {
-            throw new IllegalStateException("The Issuer \"" + metadataIssuer + "\" provided in the configuration did not "
-                + "match the requested issuer \"" + issuer + "\"");
-        }
-    }
-
     private static Map<String, Object> getConfiguration(String issuer, URI... uris) {
-        String errorMessage = "Unable to resolve the Configuration with the provided Issuer of " +
-            "\"" + issuer + "\"";
+        String errorMessage = "Unable to resolve the Configuration with the provided Issuer of " + issuer;
         for (URI uri : uris) {
             try {
                 RequestEntity<Void> request = RequestEntity.get(uri).build();
@@ -78,16 +62,5 @@ public class AADJwtDecoderProviderConfiguration {
             .build(Collections.emptyMap());
     }
 
-    private static URI oidcRfc8414(URI issuer) {
-        return UriComponentsBuilder.fromUri(issuer)
-            .replacePath(OIDC_METADATA_PATH + issuer.getPath())
-            .build(Collections.emptyMap());
-    }
-
-    private static URI oauth(URI issuer) {
-        return UriComponentsBuilder.fromUri(issuer)
-            .replacePath(OAUTH_METADATA_PATH + issuer.getPath())
-            .build(Collections.emptyMap());
-    }
 }