+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/AADTrustedIssuerRepository.java
@@ -3,10 +3,13 @@
 package com.azure.spring.aad;
 
 import com.azure.spring.autoconfigure.b2c.AADB2CProperties.UserFlows;
+import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.Collections;
+import java.util.List;
 import java.util.stream.Collectors;
+import org.apache.commons.lang.ArrayUtils;
+import org.springframework.util.Assert;
 import org.springframework.util.StringUtils;
 
 /**
@@ -17,53 +20,54 @@ public class AADTrustedIssuerRepository {
     private static final String LOGIN_MICROSOFT_ONLINE_ISSUER = "https://login.microsoftonline.com/";
     private static final String STS_WINDOWS_ISSUER = "https://sts.windows.net/";
     private static final String STS_CHINA_CLOUD_API_ISSUER = "https://sts.chinacloudapi.cn/";
-
-    private final Set<String> trustedIssuers = new HashSet<>();
+    private static final String PATH_DELIMITER = "/";
+    private static final String PATH_DELIMITER_V2 = "/v2.0";
+    private final List<String> trustedIssuers = new ArrayList<>();
+    private final String tenantId;
 
     public AADTrustedIssuerRepository(String tenantId) {
-        this(tenantId, null, null);
+        this.tenantId = tenantId;
+        trustedIssuers.addAll(buildAADIssuer(PATH_DELIMITER));
+        trustedIssuers.addAll(buildAADIssuer(PATH_DELIMITER_V2));
     }
 
-    public AADTrustedIssuerRepository(String tenantId, String tenantName, UserFlows userFlows) {
-
-        trustedIssuers.addAll(buildV1Issuer(tenantId));
-        trustedIssuers.addAll(buildV2Issuer(tenantId));
-        if (userFlows != null && !StringUtils.isEmpty(tenantName)) {
-            trustedIssuers.add(buildB2CIssuer(tenantId, tenantName));
-            trustedIssuers.addAll(buildB2CPolicyIssuer(tenantId, tenantName, userFlows));
-        }
-    }
-
-    private Set<String> buildV1Issuer(String tenantId) {
+    private List<String> buildAADIssuer(String delimiter) {
         return Arrays.asList(LOGIN_MICROSOFT_ONLINE_ISSUER, STS_WINDOWS_ISSUER, STS_CHINA_CLOUD_API_ISSUER).stream()
-            .map(s -> s + tenantId + "/").collect(Collectors.toSet());
+            .map(s -> s + tenantId + delimiter).collect(Collectors.toList());
     }
 
-    private Set<String> buildV2Issuer(String tenantId) {
-        return Arrays.asList(LOGIN_MICROSOFT_ONLINE_ISSUER, STS_WINDOWS_ISSUER, STS_CHINA_CLOUD_API_ISSUER).stream()
-            .map(s -> s + tenantId + "/v2.0").collect(Collectors.toSet());
+    public void buildB2CIssuer(String tenantName) {
+        Assert.notNull(tenantName, "tenantName cannot be null.");
+        trustedIssuers.add(String.format("https://%s.b2clogin.com/%s/v2.0/", tenantName, tenantId));
     }
 
-    private String buildB2CIssuer(String tenantId, String tenantName) {
-        return String.format("https://%s.b2clogin.com/%s/v2.0/", tenantName, tenantId);
-    }
-
-    private Set<String> buildB2CPolicyIssuer(String tenantId, String tenantName, UserFlows userFlows) {
-        Set<String> trustedPolicyIssuers = new HashSet<>();
-        trustedPolicyIssuers.add(String
+    /**
+     * Only the V2 version of Access Token is supported when using Azure AD B2C user flows.
+     */
+    public void buildB2CUserFlowsIssuer(String tenantName, UserFlows userFlows) {
+        Assert.notNull(tenantName, "tenantName cannot be null.");
+        Assert.notNull(userFlows, "userFlows cannot be null.");
+        trustedIssuers.add(String
             .format("https://%s.b2clogin.com/tfp/%s/%s/v2.0/", tenantName, tenantId, userFlows.getSignUpOrSignIn()));
         if (!StringUtils.isEmpty(userFlows.getProfileEdit())) {
-            trustedPolicyIssuers.add(String
+            trustedIssuers.add(String
                 .format("https://%s.b2clogin.com/tfp/%s/%s/v2.0/", tenantName, tenantId, userFlows.getProfileEdit()));
         }
         if (!StringUtils.isEmpty(userFlows.getPasswordReset())) {
-            trustedPolicyIssuers.add(String
+            trustedIssuers.add(String
                 .format("https://%s.b2clogin.com/tfp/%s/%s/v2.0/", tenantName, tenantId, userFlows.getPasswordReset()));
         }
-        return trustedPolicyIssuers;
     }
 
-    public Set<String> getTrustedIssuers() {
-        return trustedIssuers;
+    public List<String> getTrustedIssuers() {
+        return Collections.unmodifiableList(trustedIssuers);
     }
+
+    public boolean addTrustedIssuer(String... issuer) {
+        if(ArrayUtils.isEmpty(issuer)){
+            return false;
+        }
+        return trustedIssuers.addAll(Arrays.stream(issuer).collect(Collectors.toSet()));
+    }
+
 }