+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/validator/AADJwtAudienceValidator.java
@@ -24,12 +24,7 @@ public class AADJwtAudienceValidator implements OAuth2TokenValidator<Jwt> {
     @SuppressWarnings({"unchecked", "rawtypes"})
     public AADJwtAudienceValidator(List<String> audiences) {
         Assert.notNull(audiences, "audiences cannot be null");
-        if (!audiences.isEmpty()) {
-            this.validator = new AADJwtClaimValidator<>(AADTokenClaim.AUD,
-                aud -> audiences.containsAll(aud));
-        } else {
-            this.validator = new AADJwtClaimValidator(AADTokenClaim.AUD, aud -> true);
-        }
+        this.validator = new AADJwtClaimValidator<>(AADTokenClaim.AUD, aud -> audiences.containsAll(aud));
     }
 
     /**