+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.0 (2021-01-20)
 ### Breaking Changes
 - Exposed `userNameAttributeName` to configure the user's name.
+- Change `tenant` to `tenant-name` in the configuration file.
 
 
 ## 3.0.0 (2020-12-30)